requirements.txt
# Core dependencies
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.9.0
pydantic-settings>=2.6.0
python-multipart>=0.0.20

# ML/AI dependencies
torch>=2.0.0
transformers>=4.37.0
vllm>=0.6.0
qwen-vl-utils>=0.0.8
accelerate>=0.34.0

# Image/Video processing
Pillow>=10.0.0
decord>=0.6.0
opencv-python>=4.8.0
numpy>=1.24.0

# HTTP client
requests>=2.32.0
httpx>=0.27.0

# Logging and monitoring
python-json-logger>=2.0.7

# Development dependencies (optional)
pytest>=8.0.0
pytest-asyncio>=0.24.0
black>=24.0.0
isort>=5.13.0
flake8>=7.0.0
